Refactors goal summary into sub-component
